+++ src/AddTodoForm/AddTodoForm.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+function AddTodoForm() {
+  return (
+    <div>
+          <form>
+              <label htmlFor="todoTitle" >Title</label>  
+              <input type='text' id="todoTitle" />
+               <button type='summit'>Add </button>
+          </form>
+        
+    </div>
+  )
+}
+export default AddTodoForm ;
